softdeskAPI/pyproject.toml
[tool.poetry]
name = "softdeskapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "^3.7.2"
Django = "^4.2.4"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pytz = "^2023.3"
sqlparse = "^0.4.4"
tzdata = "^2023.3"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
